requirements.txt
mitmproxy>=10.0
pystray>=0.19
Pillow>=10.0
